Replace WeakStorageSpec with WeakStorageTests.
